feat(core-ui): move and refactor theme to multiplatform
