+++ app/build.gradle.kts
@@ -76,5 +76,6 @@ dependencies {
   debugImplementation(libs.androidx.ui.test.manifest)
 
   implementation(project(":common"))
+  implementation(project(":core:ui"))
   ksp(project(":ksp"))
 }
