+++ core/ui/build.gradle.kts
@@ -37,6 +37,7 @@ kotlin {
       }
       androidMain.dependencies {
         implementation(libs.androidx.ui.tooling.preview)
+        implementation(libs.androidx.core.ktx)
       }
     }
   }
