+++ core/ui/src/commonMain/kotlin/dev/teogor/stitch/ui/theme/Theme.kt
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2026 teogor (Teodor Grigor)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.teogor.stitch.ui.theme
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+
+internal val DarkColorScheme = darkColorScheme(
+  primary = Purple80,
+  secondary = PurpleGrey80,
+  tertiary = Pink80,
+)
+
+internal val LightColorScheme = lightColorScheme(
+  primary = Purple40,
+  secondary = PurpleGrey40,
+  tertiary = Pink40,
+)
+
+@Composable
+@Suppress("ktlint:standard:function-naming")
+fun StitchTheme(
+  darkTheme: Boolean = isSystemInDarkTheme(),
+  // Dynamic color is available on Android 12+
+  dynamicColor: Boolean = true,
+  content: @Composable () -> Unit,
+) {
+  val colorScheme = getColorScheme(darkTheme, dynamicColor)
+
+  applyPlatformTheme(colorScheme, darkTheme)
+
+  MaterialTheme(
+    colorScheme = colorScheme,
+    typography = Typography,
+    content = content,
+  )
+}
+
+@Composable
+internal expect fun getColorScheme(darkTheme: Boolean, dynamicColor: Boolean): ColorScheme
+
+@Composable
+internal expect fun applyPlatformTheme(colorScheme: ColorScheme, darkTheme: Boolean)
