+++ core/ui/src/jvmMain/kotlin/dev/teogor/stitch/ui/theme/Theme.jvm.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 teogor (Teodor Grigor)
+ * Copyright 2026 teogor (Teodor Grigor)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,15 @@
 
 package dev.teogor.stitch.ui.theme
 
-import androidx.compose.ui.graphics.Color
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.Composable
 
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
+@Composable
+internal actual fun getColorScheme(darkTheme: Boolean, dynamicColor: Boolean): ColorScheme {
+  return if (darkTheme) DarkColorScheme else LightColorScheme
+}
 
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+@Composable
+internal actual fun applyPlatformTheme(colorScheme: ColorScheme, darkTheme: Boolean) {
+  // No-op for Desktop
+}
