+++ core/ui/src/wasmJsMain/kotlin/dev/teogor/stitch/ui/theme/Theme.wasmJs.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2026 teogor (Teodor Grigor)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.teogor.stitch.ui.theme
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.Composable
+
+@Composable
+internal actual fun getColorScheme(darkTheme: Boolean, dynamicColor: Boolean): ColorScheme {
+  return if (darkTheme) DarkColorScheme else LightColorScheme
+}
+
+@Composable
+internal actual fun applyPlatformTheme(colorScheme: ColorScheme, darkTheme: Boolean) {
+  // No-op for WasmJS
+}
